add header to all page and add some design

+++ EducationManagementSystem/Screens/ExamDetails.js
@@ -1,6 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import { View, Text, FlatList, ActivityIndicator, StyleSheet, Button, Alert, TextInput } from 'react-native';
+import { View, Text, FlatList, ActivityIndicator, StyleSheet, Button, Alert, TextInput, TouchableOpacity, ScrollView} from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import CustomHeader from '../Components/CustomHeader';
+import NavigateToHomeButton from '../Components/NavigateToHomeButton';
 
 const ExamDetails = ({ route, navigation }) => {
   const { examId, email } = route.params;
@@ -65,29 +67,47 @@ const ExamDetails = ({ route, navigation }) => {
   };
 
   const handleDeleteResult = async () => {
-    try {
-      const response = await fetch(`http://192.168.136.1:3002/studentsExam/deleteResult?examId=${examId}`, {
-        method: 'DELETE',
-      });
-
-      if (!response.ok) {
-        const errorMessage = await response.text();
-        throw new Error(`Failed to delete exam: ${errorMessage}`);
-      }
-
-      Alert.alert('Success', 'Exam and associated subjects deleted successfully', [
+    // Prompt a confirmation message before deleting
+    Alert.alert(
+      'Confirm Deletion',
+      'Are you sure you want to delete this result?',
+      [
         {
-          text: 'OK',
-          onPress: () => {
-            navigation.navigate('Home');
+          text: 'Cancel',
+          style: 'cancel',
+        },
+        {
+          text: 'Delete',
+          onPress: async () => {
+            try {
+              const response = await fetch(`http://192.168.136.1:3002/studentsExam/deleteResult?examId=${examId}`, {
+                method: 'DELETE',
+              });
+  
+              if (!response.ok) {
+                const errorMessage = await response.text();
+                throw new Error(`Failed to delete exam: ${errorMessage}`);
+              }
+  
+              Alert.alert('Success', 'Exam and associated subjects deleted successfully', [
+                {
+                  text: 'OK',
+                  onPress: () => {
+                    navigation.navigate('Home');
+                  },
+                },
+              ]);
+            } catch (error) {
+              console.error('Error deleting exam:', error);
+              Alert.alert('Error', 'Failed to delete exam. Please try again.');
+            }
           },
         },
-      ]);
-    } catch (error) {
-      console.error('Error deleting exam:', error);
-      Alert.alert('Error', 'Failed to delete exam. Please try again.');
-    }
+      ],
+      { cancelable: false }
+    );
   };
+  
 
   const formatExamDate = (examDate) => {
     const date = new Date(examDate);
@@ -161,8 +181,12 @@ const ExamDetails = ({ route, navigation }) => {
   const totalMarks = examResults.slice(1).reduce((total, row) => total + parseInt(row.Mark, 10) || 0, 0);
   const totalSubjects = examResults.length - 1;
   const averageMarks = totalSubjects > 0 ? totalMarks / totalSubjects : 0;
-  return (
-    <View>
+
+return (
+  <View>
+    <CustomHeader />
+    <ScrollView>
+      
       {isLoading ? (
         <ActivityIndicator size="large" color="#0000ff" />
       ) : error ? (
@@ -171,63 +195,81 @@ const ExamDetails = ({ route, navigation }) => {
         </View>
       ) : (
         <>
-          <Text>Student Email: {email}</Text>
-          <Text>Exam Name: {examDetails.ExamName}</Text>
-          <Text>Exam Date: {formatExamDate(examDetails.ExamDate)}</Text>
-  
-          <View style={{ borderStyle: 'solid', borderWidth: 1, borderColor: 'gray', marginBottom: 10 }}>
-            <View style={{ flexDirection: 'row', marginVertical: 5 }}>
-              <Text style={{ flex: 1, padding: 5, fontWeight: 'bold' }}>Subject ID</Text>
-              <Text style={{ flex: 1, padding: 5, fontWeight: 'bold' }}>Subject Name</Text>
-              <Text style={{ flex: 1, padding: 5, fontWeight: 'bold' }}>Mark</Text>
+          <View style={styles.infoContainer}>
+            <Text style={styles.infoText}>Student Email: {email}</Text>
+            <Text style={styles.infoText}>Exam Name: {examDetails.ExamName}</Text>
+            <Text style={styles.infoText}>Exam Date: {formatExamDate(examDetails.ExamDate)}</Text>
+          </View>
+
+          <View style={styles.container}>
+            <View style={styles.tableHeader}>
+              <Text style={styles.headerText}>Subject ID</Text>
+              <Text style={styles.headerText}>Subject Name</Text>
+              <Text style={styles.headerText}>Mark</Text>
             </View>
   
             <FlatList
               data={examResults}
               renderItem={({ item, index }) => (
-                <View style={{ flexDirection: 'row', marginVertical: 5 }}>
-                  <Text style={{ flex: 1, padding: 5 }}>{item.SubjectID}</Text>
-                  <Text style={{ flex: 1, padding: 5 }}>{item.SubjectName}</Text>
-                  <View style={{ flex: 1, padding: 5 }}>
+                <View style={styles.row}>
+                  <Text style={styles.cell}>{item.SubjectID}</Text>
+                  <Text style={styles.cell}>{item.SubjectName}</Text>
                     {isEditMode ? (
                       <TextInput
                         defaultValue={editedMarks[index] !== undefined ? editedMarks[index] : item.Mark.toString()}
                         onChangeText={(text) => handleEditMark(index, text)}
-                        style={{
-                          height: 40,
-                          borderColor: 'gray',
-                          borderWidth: 1,
-                          borderRadius: 5,
-                          padding: 5,
-                          backgroundColor: '#F0F8FF', // Light blue background color
-                        }}
+                        style={styles.editInput}
                         placeholder="Edit Mark"
                       />
                     ) : (
-                      <Text>{item.Mark}</Text>
+                      <Text style={styles.cell}>{item.Mark}</Text>
                     )}
-                  </View>
                 </View>
               )}
               keyExtractor={(item, index) => index.toString()}
             />
+  
+            <View style={styles.line} />
+  
+            <View style={styles.averageContainer}>
+              <Text style={styles.averageLabel}>Average Marks:</Text>
+              <Text style={styles.averageValue}>{averageMarks.toFixed(2)}</Text>
+            </View>
           </View>
   
-          <Text>Average Marks: {averageMarks.toFixed(2)}</Text>
-  
-          <View style={styles.buttonContainer}>
-            {!userData || (userData.student !== 1 && (
-              <Button title={isEditMode ? 'Save Changes' : 'Edit Result'} onPress={isEditMode ? handleSaveChanges : handleEditResult} />
-            ))}
-            {!userData || (userData.student !== 1 && (
-              <Button title="Delete Result" onPress={handleDeleteResult} />
-            ))}
-          </View>
+          {userData.admin === 1 || userData.teacher === 1 ? (
+            !isEditMode && (
+              <View style={styles.buttonContainer}>
+                <TouchableOpacity style={styles.button} onPress={handleEditResult}>
+                  <Text style={styles.buttonText}>Edit Result</Text>
+                </TouchableOpacity>
+                <TouchableOpacity style={styles.button} onPress={handleDeleteResult}>
+                  <Text style={styles.buttonText}>Delete Result</Text>
+                </TouchableOpacity>
+              </View>
+            )
+          ) : (
+            userData.student === 1 ? (
+              <View style={{ alignItems: 'flex-end', marginRight: 20 }}>
+                <NavigateToHomeButton />
+              </View>
+            ) : null
+          )}
+
+          {isEditMode && (
+            <View style={[styles.buttonContainer, { justifyContent: 'flex-end' }]}>
+              <TouchableOpacity style={styles.button} onPress={handleSaveChanges}>
+                <Text style={styles.buttonText}>Save Changes</Text>
+              </TouchableOpacity>
+            </View>
+          )}
         </>
       )}
-    </View>
+    </ScrollView>
+  </View>
   );
-  
+
+
 };
 
 const styles = StyleSheet.create({
@@ -241,6 +283,102 @@ const styles = StyleSheet.create({
     justifyContent: 'space-between',
     margin: 10,
   },
+  button: {
+    backgroundColor: '#4494ad',
+    paddingVertical: 12,
+    paddingHorizontal: 20,
+    borderRadius: 8,
+    elevation: 5,
+    marginRight:10,
+    marginLeft:10,
+    marginBottom:100,
+  },
+  container: {
+    backgroundColor: '#fff',
+    borderRadius: 10,
+    padding: 20,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+    elevation: 5,
+    margin: 20,
+  },
+  tableHeader: {
+    flexDirection: 'row',
+    borderBottomWidth: 1,
+    borderBottomColor: '#000',
+    paddingHorizontal: 10,
+  },
+  headerText: {
+    flex: 1,
+    paddingVertical: 8,
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+  row: {
+    flexDirection: 'row',
+    marginVertical: 10,
+    paddingHorizontal: 10,
+    alignItems: 'center',
+  },
+  cell: {
+    flex: 1,
+    paddingVertical: 8,
+    textAlign: 'center',
+  },
+  editInput: {
+    height: 40,
+    borderColor: 'gray',
+    borderWidth: 1,
+    borderRadius: 5,
+    padding: 10,
+    backgroundColor: '#F0F8FF',
+    textAlign: 'center',
+  },
+  line: {
+    borderBottomWidth: 1,
+    borderBottomColor: 'black',
+    marginBottom: 10,
+  },
+  averageContainer: {
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  averageLabel: {
+    fontWeight: 'bold',
+  },
+  averageValue: {
+    marginLeft: 5,
+  },
+  buttonText: {
+    color: 'white',
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+  infoContainer: {
+    backgroundColor: '#fff',
+    borderRadius: 10,
+    padding: 20,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+    elevation: 5,
+    margin: 20,
+  },
+  infoText: {
+    marginBottom: 10,
+    fontWeight: 'bold',
+  },
+  
 });
 
 export default ExamDetails;
